+++ ghe-jira-connector/static/.gitignore
