--- ghe-jira-connector/static/admin-page/src/index.html
@@ -1,17 +1,0 @@
-<html>
-<head>
-    <title>
-        Our APP
-    </title>
-
-</head>
-<body>
-<div id="root">
-
-
-</div>
-
-
-<script src="/bundle.js"></script>
-</body>
-</html>
