+++ ghe-jira-connector/static/package.json
@@ -36,6 +36,7 @@
     "@forge/babel-plugin-transform-ui": "^1.0.0",
     "@forge/bridge": "0.1.0-next.3",
     "@types/react-dom": "^17.0.9",
+    "@types/styled-components": "^5.1.14",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.3",
