+++ ghe-jira-connector/static/src/README.md
@@ -1,0 +1,1 @@
+Where we should put all frontend specific things.
