+++ ghe-jira-connector/static/src/common/copy.ts
@@ -1,0 +1,3 @@
+export const jiraAdminPageCopy = {
+  emptyStateTitle: "Connect GitHub Enterprise to Jira Software"
+};
