+++ ghe-jira-connector/static/src/domain/WebhookUrlResponse.tsx
