+++ ghe-jira-connector/static/src/index.html
@@ -1,0 +1,10 @@
+<html>
+  <head>
+    <title>GitHub Enterprise Jira Connector</title>
+  </head>
+  <body>
+    <div id="root"></div>
+
+    <script src="/bundle.js"></script>
+  </body>
+</html>
