+++ ghe-jira-connector/static/src/index.styles.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { fontFamily as akFontFamily } from "@atlaskit/theme";
+
+export const Container = styled.div`
+  font-family: ${akFontFamily()};
+`;
