+++ ghe-jira-connector/static/src/resources/base64-images.ts
