+++ ghe-jira-connector/static/src/ui/admin-page/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { ContentTitle } from './AdminPage.styles';
+import { jiraAdminPageCopy } from '../../common/copy';
+
+export const AdminPage = () => {
+    return (
+        <>
+            <ContentTitle>{jiraAdminPageCopy.emptyStateTitle}</ContentTitle>
+        </>
+    )
+};
