+++ ghe-jira-connector/static/tsconfig.json
@@ -19,6 +19,6 @@
     "jsx": "react"
   },
   "include": [
-    "src"
+    "./src"
   ]
 }
