+++ ghe-jira-connector/static/webpack.config.ts
@@ -67,6 +67,7 @@ const config: webpack.Configuration = {
     resolve: {
         extensions: ['.tsx', '.ts', '.js'],
     },
+
 };
 
 export default config;
